+++ GEN_QUIDGEST/GQT/GenioMVC/Models/Ccorr.cs
@@ -38,7 +38,7 @@ namespace GenioMVC.Models
 		[DisplayName("Order")]
 		/// <summary>Field : "Order" Tipo: "N" Formula:  ""</summary>
 		[NumericAttribute(0)]
-		public decimal? ValNorder { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValNorder, 0)); } set { klass.ValNorder = Convert.ToDouble(value); } }
+		public decimal? ValNorder { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValNorder, 0)); } set { klass.ValNorder = Convert.ToDecimal(value); } }
 		public bool ShouldSerializeValNorder() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Ccorr.ValNorder");
 
 		[DisplayName("Instant")]
@@ -74,13 +74,13 @@ namespace GenioMVC.Models
 		[DisplayName("Amount")]
 		/// <summary>Field : "Amount" Tipo: "N" Formula:  ""</summary>
 		[NumericAttribute(0)]
-		public decimal? ValQnty { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValQnty, 0)); } set { klass.ValQnty = Convert.ToDouble(value); } }
+		public decimal? ValQnty { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValQnty, 0)); } set { klass.ValQnty = Convert.ToDecimal(value); } }
 		public bool ShouldSerializeValQnty() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Ccorr.ValQnty");
 
 		[DisplayName("Balance")]
 		/// <summary>Field : "Balance" Tipo: "N" Formula:  ""</summary>
 		[NumericAttribute(0)]
-		public decimal? ValBalance { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValBalance, 0)); } set { klass.ValBalance = Convert.ToDouble(value); } }
+		public decimal? ValBalance { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValBalance, 0)); } set { klass.ValBalance = Convert.ToDecimal(value); } }
 		public bool ShouldSerializeValBalance() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Ccorr.ValBalance");
 
 		[DisplayName("Ref")]
