+++ GEN_QUIDGEST/GQT/GenioMVC/Models/Dilin.cs
@@ -47,7 +47,7 @@ namespace GenioMVC.Models
 		[DisplayName("Line")]
 		/// <summary>Field : "Line" Tipo: "N" Formula:  ""</summary>
 		[NumericAttribute(0)]
-		public decimal? ValLinenumb { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValLinenumb, 0)); } set { klass.ValLinenumb = Convert.ToDouble(value); } }
+		public decimal? ValLinenumb { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValLinenumb, 0)); } set { klass.ValLinenumb = Convert.ToDecimal(value); } }
 		public bool ShouldSerializeValLinenumb() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Dilin.ValLinenumb");
 
 		[DisplayName(">>PRODUCT")]
@@ -62,19 +62,19 @@ namespace GenioMVC.Models
 		[DisplayName("Ordered")]
 		/// <summary>Field : "Ordered" Tipo: "N" Formula:  ""</summary>
 		[NumericAttribute(0)]
-		public decimal? ValOrdered { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValOrdered, 0)); } set { klass.ValOrdered = Convert.ToDouble(value); } }
+		public decimal? ValOrdered { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValOrdered, 0)); } set { klass.ValOrdered = Convert.ToDecimal(value); } }
 		public bool ShouldSerializeValOrdered() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Dilin.ValOrdered");
 
 		[DisplayName("Delivered")]
 		/// <summary>Field : "Delivered" Tipo: "N" Formula:  ""</summary>
 		[NumericAttribute(0)]
-		public decimal? ValDelivere { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValDelivere, 0)); } set { klass.ValDelivere = Convert.ToDouble(value); } }
+		public decimal? ValDelivere { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValDelivere, 0)); } set { klass.ValDelivere = Convert.ToDecimal(value); } }
 		public bool ShouldSerializeValDelivere() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Dilin.ValDelivere");
 
 		[DisplayName("Outstanding")]
 		/// <summary>Field : "Outstanding" Tipo: "N" Formula: + "[DILIN->ORDERED]-[DILIN->DELIVERE]"</summary>
 		[NumericAttribute(0)]
-		public decimal? ValOutstand { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValOutstand, 0)); } set { klass.ValOutstand = Convert.ToDouble(value); } }
+		public decimal? ValOutstand { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValOutstand, 0)); } set { klass.ValOutstand = Convert.ToDecimal(value); } }
 		public bool ShouldSerializeValOutstand() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Dilin.ValOutstand");
 
 		[DisplayName("Instant")]
