+++ GEN_QUIDGEST/GQT/GenioMVC/Models/Dttyp.cs
@@ -74,61 +74,61 @@ namespace GenioMVC.Models
 
 		[DisplayName("Conditional (smallint) (storage: 2 byte)")]
 		/// <summary>Field : "Conditional (smallint) (storage: 2 byte)" Tipo: "IF" Formula:  ""</summary>
-		public double ValBoolean2 { get { return klass.ValBoolean2; } set { klass.ValBoolean2 = value; } }
+		public decimal ValBoolean2 { get { return klass.ValBoolean2; } set { klass.ValBoolean2 = value; } }
 		public bool ShouldSerializeValBoolean2() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Dttyp.ValBoolean2");
 
 		[DisplayName("Numeric  4.0 - small integer (storage: 2 byte)")]
 		/// <summary>Field : "Numeric  4.0 - small integer (storage: 2 byte)" Tipo: "N" Formula:  ""</summary>
 		[NumericAttribute(0)]
-		public decimal? ValSmallint { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValSmallint, 0)); } set { klass.ValSmallint = Convert.ToDouble(value); } }
+		public decimal? ValSmallint { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValSmallint, 0)); } set { klass.ValSmallint = Convert.ToDecimal(value); } }
 		public bool ShouldSerializeValSmallint() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Dttyp.ValSmallint");
 
 		[DisplayName("Numeric  9.0 - integer (storage: 4 byte)")]
 		/// <summary>Field : "Numeric  9.0 - integer (storage: 4 byte)" Tipo: "N" Formula:  ""</summary>
 		[NumericAttribute(0)]
-		public decimal? ValInteger { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValInteger, 0)); } set { klass.ValInteger = Convert.ToDouble(value); } }
+		public decimal? ValInteger { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValInteger, 0)); } set { klass.ValInteger = Convert.ToDecimal(value); } }
 		public bool ShouldSerializeValInteger() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Dttyp.ValInteger");
 
 		[DisplayName("Numeric 15.0 - big integer (storage: 8 byte)")]
 		/// <summary>Field : "Numeric 15.0 - big integer (storage: 8 byte)" Tipo: "N" Formula:  ""</summary>
 		[NumericAttribute(0)]
-		public decimal? ValBigint { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValBigint, 0)); } set { klass.ValBigint = Convert.ToDouble(value); } }
+		public decimal? ValBigint { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValBigint, 0)); } set { klass.ValBigint = Convert.ToDecimal(value); } }
 		public bool ShouldSerializeValBigint() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Dttyp.ValBigint");
 
 		[DisplayName("Numeric  8.2 real=float(24) (precision 7 digits) (storage: 4 byte)")]
 		/// <summary>Field : "Numeric  8.2 real=float(24) (precision 7 digits) (storage: 4 byte)" Tipo: "N" Formula:  ""</summary>
 		[NumericAttribute(2)]
-		public decimal? ValReal { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValReal, 2)); } set { klass.ValReal = Convert.ToDouble(value); } }
+		public decimal? ValReal { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValReal, 2)); } set { klass.ValReal = Convert.ToDecimal(value); } }
 		public bool ShouldSerializeValReal() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Dttyp.ValReal");
 
 		[DisplayName("Numeric 15.2 double = float(53) (precision 15 digits) (storage: 8 byte)")]
 		/// <summary>Field : "Numeric 15.2 double = float(53) (precision 15 digits) (storage: 8 byte)" Tipo: "N" Formula:  ""</summary>
 		[NumericAttribute(2)]
-		public decimal? ValFloat { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValFloat, 2)); } set { klass.ValFloat = Convert.ToDouble(value); } }
+		public decimal? ValFloat { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValFloat, 2)); } set { klass.ValFloat = Convert.ToDecimal(value); } }
 		public bool ShouldSerializeValFloat() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Dttyp.ValFloat");
 
 		[DisplayName("Decimal (1-10) (storage: 5 byte)")]
 		/// <summary>Field : "Decimal (1-10) (storage: 5 byte)" Tipo: "ND" Formula:  ""</summary>
 		[NumericAttribute(4)]
-		public decimal? ValDecimal { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValDecimal, 4)); } set { klass.ValDecimal = Convert.ToDouble(value); } }
+		public decimal? ValDecimal { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValDecimal, 4)); } set { klass.ValDecimal = Convert.ToDecimal(value); } }
 		public bool ShouldSerializeValDecimal() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Dttyp.ValDecimal");
 
 		[DisplayName("Decimal (11-15) (storage: 9 byte)")]
 		/// <summary>Field : "Decimal (11-15) (storage: 9 byte)" Tipo: "ND" Formula:  ""</summary>
 		[NumericAttribute(4)]
-		public decimal? ValDecimal9 { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValDecimal9, 4)); } set { klass.ValDecimal9 = Convert.ToDouble(value); } }
+		public decimal? ValDecimal9 { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValDecimal9, 4)); } set { klass.ValDecimal9 = Convert.ToDecimal(value); } }
 		public bool ShouldSerializeValDecimal9() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Dttyp.ValDecimal9");
 
 		[DisplayName("Money - decimal (1-10) (storage: 5 byte)")]
 		/// <summary>Field : "Money - decimal (1-10) (storage: 5 byte)" Tipo: "$D" Formula:  ""</summary>
 		[CurrencyAttribute("EUR", 4)]
-		public decimal? ValMoney { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValMoney, 4)); } set { klass.ValMoney = Convert.ToDouble(value); } }
+		public decimal? ValMoney { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValMoney, 4)); } set { klass.ValMoney = Convert.ToDecimal(value); } }
 		public bool ShouldSerializeValMoney() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Dttyp.ValMoney");
 
 		[DisplayName("Money - decimal (11-15) (storage: 9 byte)")]
 		/// <summary>Field : "Money - decimal (11-15) (storage: 9 byte)" Tipo: "$D" Formula:  ""</summary>
 		[CurrencyAttribute("EUR", 4)]
-		public decimal? ValMoney9 { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValMoney9, 4)); } set { klass.ValMoney9 = Convert.ToDouble(value); } }
+		public decimal? ValMoney9 { get { return Convert.ToDecimal(GlobalFunctions.RoundQG(klass.ValMoney9, 4)); } set { klass.ValMoney9 = Convert.ToDecimal(value); } }
 		public bool ShouldSerializeValMoney9() => this.SerializeAllFields || this.FieldsToSerialize.Contains("Dttyp.ValMoney9");
 
 		[DisplayName("Date")]
